+++ constant/tasks/tasks_name_constants.py
@@ -1,1 +1,1 @@
-TASK_GET_OFFERS_AMAZON = "amazon_api.get_category_offers"
+TASK_GET_OFFERS_AMAZON = "amazon_api_task.get_category_offers_task"
